backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
chispa==0.10.1
click==8.1.8
email-validator==2.0.0
fastapi==0.95.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
numpy==1.24.3
packaging==24.2
pandas==2.0.1
passlib==1.7.4
pluggy==1.5.0
prettytable==3.12.0
py4j==0.10.9.7
pyarrow==18.1.0
pydantic==1.10.7
pydantic_core==2.27.2
PyJWT==2.6.0
pyspark==3.5.4
pytest==7.3.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.21.1
wcwidth==0.2.13
bcrypt==4.0.1
python-dotenv==1.0.0
supabase==1.0.3
python-multipart==0.0.6
requests==2.28.2
python-jose==3.3.0
httpx==0.24.0
prefect==2.10.21
asyncio==3.4.3
aiohttp==3.8.4
httptools==0.5.0
aiosqlite==0.19.0
